requirements.txt
streamlit==1.42.2
pandas==2.2.3
python-dotenv==1.0.1
langchain==0.3.19
langchain-groq==0.2.4
python-decouple==3.8